requirements.txt
Flask
azure-identity
azure-cosmos
azure-storage-blob
